# Add support for cachestat() syscall